Modernize to 10.14 SDK.

+++ TalkingMoose/UKRecordedSpeechChannel.m
@@ -69,7 +69,7 @@
 
 -(void)	spendTime: (NSTimer*)tim
 {
-	int numPhonemeEntries = [phonemes count];
+	NSUInteger numPhonemeEntries = [phonemes count];
 	if( !phonemes || currListEntryIdx < 0 || currListEntryIdx >= numPhonemeEntries )
 	{
 		currListEntryIdx = -1;
